chore(DataSpreadsheet): create shared util for active cell header check
