+++ packages/cloud-cognitive/src/components/DataSpreadsheet/DataSpreadsheetBody.js
@@ -14,6 +14,7 @@ import { pkg } from '../../settings';
 import { deepCloneObject } from '../../global/js/utils/deepCloneObject';
 import uuidv4 from '../../global/js/utils/uuidv4';
 import { createCellSelectionArea } from './createCellSelectionArea';
+import { checkActiveHeaderCell } from './checkActiveHeaderCell';
 const blockClass = `${pkg.prefix}--data-spreadsheet`;
 
 export const DataSpreadsheetBody = forwardRef(
@@ -253,32 +254,6 @@ export const DataSpreadsheetBody = forwardRef(
       [clickAndHoldActive, currentMatcher, setSelectionAreas]
     );
 
-    const checkSelectionAreaForActiveHeader = useCallback(
-      (headerIndex) => {
-        // Determines if a row header cell should receive a highlight/active background color
-        // Check each object in selectionAreas and see if the headerIndex is between
-        // point1.row and point2.row, inclusive
-        const areasCloned = deepCloneObject(selectionAreas);
-        const activeRowIndexes = [];
-        areasCloned.forEach((area) => {
-          const greatestRowIndex = Math.max(area.point1?.row, area.point2?.row);
-          const lowestRowIndex = Math.min(area.point1?.row, area.point2?.row);
-          for (let i = lowestRowIndex; i <= greatestRowIndex; i++) {
-            activeRowIndexes.push(i);
-          }
-        });
-        const activeRowIndexesNoDuplicates = [...new Set(activeRowIndexes)];
-        if (
-          areasCloned?.length &&
-          activeRowIndexesNoDuplicates.includes(headerIndex)
-        ) {
-          return true;
-        }
-        return false;
-      },
-      [selectionAreas]
-    );
-
     // Renders each cell in the spreadsheet body
     const RenderRow = useCallback(
       ({ index, style }) => {
@@ -303,7 +278,7 @@ export const DataSpreadsheetBody = forwardRef(
                 {
                   [`${blockClass}__td-th--active-header`]:
                     activeCellCoordinates?.row === index ||
-                    checkSelectionAreaForActiveHeader(index),
+                    checkActiveHeaderCell(index, selectionAreas, 'row'),
                 }
               )}
               style={{
@@ -343,7 +318,7 @@ export const DataSpreadsheetBody = forwardRef(
         handleBodyCellClick,
         handleBodyCellHover,
         activeCellCoordinates?.row,
-        checkSelectionAreaForActiveHeader,
+        selectionAreas,
       ]
     );
